// backend/src/types.ts
export type ArtifactID = string;

export type ArtifactKind = string;

export type ArtifactMetricGeneration = {
  eventSummaryID: EventSummaryID;
  recipeRunID: RecipeRunID;
  generationID: GenerationID;
  modelID: string;
  modelParams?: EvaluationModelParameters | undefined;
  endTimestamp: ISO8601PreciseTimestamp;
  didComplete: boolean;
};

export type ArtifactPathComponent = {
  kind?: (ArtifactKind | null) | undefined;
  id: ArtifactID;
};

export type ArtifactPathPatternComponent =
  | {
      kind: ArtifactKind;
      id?: (ArtifactID | null) | undefined;
    }
  | {
      id: ArtifactID;
    };

export type ArtifactPathPattern = ArtifactPathPatternComponent[];

export type ArtifactPathRecipeTrigger = {
  id: RecipeTriggerID;
  evaluationGroupID: EvaluationGroupID;
  name: string;
  kind: typeof RecipeTriggerKind.artifactPath;
  creationTimestamp: ISO8601PreciseTimestamp;
  updateTimestamp: ISO8601PreciseTimestamp;
  artifactPathPattern: ArtifactPathPattern;
};

export type ArtifactPath = ArtifactPathComponent[];

export type ArtifactPatternConfiguration = {
  pattern: ArtifactPathPattern;
  displayName: DisplayName;
  allowCreation?: boolean | undefined;
};
// ...
export type DisplayName = {
  zero?: string | undefined;
  one: string;
  two?: string | undefined;
  few?: string | undefined;
  many?: string | undefined;
  other: string;
  evaluate?: string | undefined;
};
// ...
export type OrgConfigurationResponse = {
  artifactPatterns: ArtifactPatternConfiguration[];
  genericArtifactName: DisplayName;
};

export type OrgEvaluationModelsRequest = {
  orgID: OrganizationID;
};

export type OrgEvaluationModelsResponse = {
  evaluationModels: EvaluationModelConfiguration[];
  defaultEvaluationModelID: string;
};

export type OrgUsersRequest = {
  orgID: OrganizationID;
};

export type OrgUsersResponse = {
  users: User[];
};

export type OrganizationID = UUID;

export type Organization = {
  id: OrganizationID;
  name: string;
  template?: OrganizationTemplate | undefined;
  isDeleted?: true | undefined;
};

export const OrganizationTemplate = {
  demo: "demo",
  general: "general",
} as const;
export type OrganizationTemplate = (typeof OrganizationTemplate)[keyof typeof OrganizationTemplate];
// ...
export type UUID = string;